requirements.txt
fastapi
uvicorn
requests
Pillow
# RPi.GPIO
# hx711
python-dotenv
SQLAlchemy